pyproject.toml
[project]
name = "aws-stepfunctions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.232.2",
    "boto3>=1.42.9",
    "botocore[crt]>=1.42.9",
    "constructs>=10.4.4",
]
